CMake: split options and conditions, closes #716
As described in the issue, the CMake process force off a cache variable
when a condition is not met. If the user installs the requirements and
re-run CMake it must also set the variable to on which is inconvenient.
The new process is to add a secondary variable and do not touch user
defined options.