Misc: rework networking
The network_stream and irc classes do not assume that owner is alive anymore by
keeping handlers before end of block. Instead, callers postpone deletion of
themselves when required to allow handler finishing correctly.

Capture all exceptions that can happen in network_stream to make sure handler is
called as appropriate in any case. Do the same in irc class.

Create a dedicated on_disconnect event in server class which is emitted when the
server gets disconnected but is not dead yet.