Irccd: do not own a queue in irc::connection class, closes #791 @1h
Since the server is an automatic state machine that recreate a new session after
a connection failure, we may have some invalid references in the irc::connection
class if there are still handler running.

Because server is already a std::shared_ptr that manages itself, it was heavy to
do the same in irc::connection class. Therefore, the irc::connection class is
now only responsible of connecting/receiving/sending data (and parsing it).

Update the server class to use a send queue and allow the user sending command
at any time which will be postponed once connected.